requirements.txt
streamlit==1.39.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
scikit-learn==1.5.2
sentence-transformers==3.0.1
spacy==3.7.4
joblib==1.4.2
PyMuPDF==1.24.9
python-docx==1.1.2
nltk==3.9.1
